@@ -3,191 +3,174 @@
 #include <Common/md5.h>
 
 #include <algorithm>
+#include <filesystem>
 #include <fstream>
-#include <string>
-#include <iostream>
-#include <memory>
 #include <utility>
 
-#define SIGN_SALT "XGRlBW9FXlekgbPrRHuSiA"
-
-namespace yandex_music 
+#include <fmt/ranges.h>
+namespace YandexMusic
 {
-    Track::Track(
-        std::string id_,
-        std::string title_,
-        std::vector<std::string> artists_,
-        std::vector<int> albums_,
-        bool available_,
-        bool lyrics_available_)
-        : id(std::move(id_))
-        , title(std::move(title_))
-        , artists(std::move(artists_))
-        , albums(std::move(albums_))
-        , available(available_)
-        , lyrics_available(lyrics_available_)
-        , supplement() 
-    {}
 
-
-    std::string Track::getTitle() const 
+    namespace
     {
-        return title;
+        constexpr std::string_view SIGN_SALT = "XGRlBW9FXlekgbPrRHuSiA";
     }
 
-
-    std::string Track::getId() const 
+    Track::Track(std::string id, std::string title,
+                 std::vector<std::string> artists,
+                 std::vector<int> albums,
+                 bool available,
+                 bool lyricsAvailable)
+        : id_(std::move(id)), title_(std::move(title)), artists_(std::move(artists)), albums_(std::move(albums)), available_(available), lyricsAvailable_(lyricsAvailable)
     {
-        return id;
     }
 
-    bool Track::getAvailable() const 
+    void Track::fetchDownloadInfo()
     {
-        return available;
+        Request request;
+        rapidjson::Document document;
+        const std::string url = "tracks/" + id_ + "/download-info";
+
+        request.makeRequest(url, document);
+        const auto &result = document["result"];
+
+        if (!result.IsArray())
+        {
+            throw std::runtime_error("Invalid response format for download info");
+        }
+
+        for (const auto &item : result.GetArray())
+        {
+            if (std::string_view{item["codec"].GetString()} == "mp3")
+            {
+                downloadInfo_.emplace(
+                    item["downloadInfoUrl"].GetString(),
+                    std::to_string(item["bitrateInKbps"].GetInt()));
+                break;
+            }
+        }
     }
 
-    bool Track::getLyricsAvailable() const 
+    void Track::fetchSupplement()
     {
-        return lyrics_available;
+        if (!lyricsAvailable_)
+            return;
+
+        Request request;
+        rapidjson::Document document;
+        const std::string url = "tracks/" + id_ + "/supplement";
+
+        request.makeRequest(url, document);
+        const auto &result = document["result"];
+
+        if (!result.IsObject())
+        {
+            throw std::runtime_error("Invalid response format for supplement");
+        }
+
+        const auto lyricsIt = result.FindMember("lyrics");
+        if (lyricsIt == result.MemberEnd())
+            return;
+
+        const auto &lyricsObj = lyricsIt->value;
+        Supplement supp;
+
+        for (const auto &field : lyricsObj.GetObject())
+        {
+            const std::string_view name{field.name.GetString()};
+
+            if (name == "id")
+            {
+                supp.setId(field.value.GetInt());
+            }
+            else if (name == "lyrics")
+            {
+                supp.setLyrics(field.value.GetString());
+            }
+            else if (name == "fullLyrics")
+            {
+                supp.setFullLyrics(field.value.GetString());
+            }
+            else if (name == "textLanguage")
+            {
+                supp.setTextLanguage(field.value.GetString());
+            }
+            else if (name == "showTranslation")
+            {
+                supp.setShowTranslation(field.value.GetBool());
+            }
+            else if (name == "hasRights")
+            {
+                supp.setHasRights(field.value.GetBool());
+            }
+        }
+
+        supplement_.emplace(std::move(supp));
     }
 
-    std::vector<std::string> Track::getArtists() const 
+    std::string Track::generateDownloadUrl(const tinyxml2::XMLDocument &xmlResponse)
     {
-        return artists;
-    }
+        const auto *downloadInfo = xmlResponse.FirstChildElement("download-info");
+        if (!downloadInfo)
+        {
+            throw std::runtime_error("Invalid XML response - missing download-info");
+        }
 
-    std::vector<int> Track::getAlbums() const 
-    {
-        return albums;
-    }
-
-    void Track::getSign(std::string &download_url, const tinyxml2::XMLDocument &xml_response) 
-    {
-        /// Parse XML
-        std::string host, path, ts, s;
-        host = xml_response.FirstChildElement("download-info")->FirstChildElement("host")->GetText();
-        path = xml_response.FirstChildElement("download-info")->FirstChildElement("path")->GetText();
-        ts = xml_response.FirstChildElement("download-info")->FirstChildElement("ts")->GetText();
-        s = xml_response.FirstChildElement("download-info")->FirstChildElement("s")->GetText();
+        const std::string host = downloadInfo->FirstChildElement("host")->GetText();
+        const std::string path = downloadInfo->FirstChildElement("path")->GetText();
+        const std::string ts = downloadInfo->FirstChildElement("ts")->GetText();
+        const std::string s = downloadInfo->FirstChildElement("s")->GetText();
 
         MD5 md5;
-        std::string text = SIGN_SALT + path.substr(1) + s;
+        const std::string text = std::string{SIGN_SALT} + path.substr(1) + s;
         md5.add(text.c_str(), text.size());
 
-        std::string sign = md5.getHash();
-
-        /// Return value
-        download_url = "https://" + host + "/" + "get-mp3/" + sign + "/" + ts + path;
+        return fmt::format("https://{}/get-mp3/{}/{}{}", host, md5.getHash(), ts, path);
     }
 
+    void Track::download(std::string_view lyricsDir, std::string_view tracksDir) const
+    {
+        if (!available_ || !downloadInfo_)
+            return;
 
-    void Track::downloadTrack(string &lyrics_dir, string &tracks_dir) {
-        /// Get track file name on fs
-        std::string name = artists.empty() ? title : artists[0] + " - " + title;
-        name = name.size() > 100 ? name.substr(0, 100) + ".mp3" : name + ".mp3";
-        std::replace(name.begin(), name.end(), '/', '-');
-
-        /// Get track download info
-        getDownloadInfo();
-        std::string url = download_info->getDownloadInfoUrl();
-        std::string bitrate = download_info->getBitrateInKbps();
-
-        tinyxml2::XMLDocument xml_response;
-        Request request;
-        request.makeRequest(url, xml_response);
-
-        std::string download_url;
-        getSign(download_url, xml_response);
-
-        /// Download track
-        Curl::get(download_url);
-        FILE *fp;
-        fp = fopen((tracks_dir + "/" + name).data(), "wb");
-        fwrite(Curl::buffer.data(), Curl::buffer.size(), 1, fp);
-        fclose(fp);
-
-        /// Download supplement
-        if (getLyricsAvailable()) 
+        // Generate filename
+        std::string filename = artists_.empty() ? title_ : artists_[0] + " - " + title_;
+        if (filename.size() > 100)
         {
-            getSupplement();
-            std::ofstream output_file(lyrics_dir + "/" + name + ".txt");
-            output_file << supplement->getFullLyrics();
+            filename.resize(100);
+        }
+        filename += ".mp3";
+        std::replace(filename.begin(), filename.end(), '/', '-');
+
+        // Download track
+        tinyxml2::XMLDocument xmlResponse;
+        Request request;
+        request.makeRequest(downloadInfo_->downloadInfoUrl(), xmlResponse);
+
+        const std::string downloadUrl = generateDownloadUrl(xmlResponse);
+        Curl::get(downloadUrl);
+
+        // Save track
+        const std::string trackPath = std::string{tracksDir} + "/" + filename;
+        std::ofstream trackFile(trackPath, std::ios::binary);
+        trackFile.write(Curl::buffer.data(), Curl::buffer.size());
+
+        // Save lyrics if available
+        if (lyricsAvailable_ && supplement_)
+        {
+            const std::string lyricsPath = std::string{lyricsDir} + "/" + filename + ".txt";
+            std::ofstream lyricsFile(lyricsPath);
+            lyricsFile << supplement_->fullLyrics();
         }
     }
 
-    void Track::getDownloadInfo() 
+    void Track::print() const
     {
-        download_info = make_shared<DownloadInfo>();
-        rapidjson::Document document;
-        Request request;
-        std::string url{"tracks/" + id + "/download-info"};
-        request.makeRequest(url, document);
-        const rapidjson::Value &result = document["result"];
-        assert(result.IsArray());
-        for (rapidjson::SizeType i = 0; i < result.Size(); i++) // Uses SizeType instead of size_t
-        {
-            string k = result[i]["codec"].GetString();
-            if (k == "mp3") 
-            {
-                download_info->setDownloadInfoUrl(result[i]["downloadInfoUrl"].GetString());
-                download_info->setBitrateInKbps(to_string(result[i]["bitrateInKbps"].GetInt()));
-            }
-        }
+        fmt::print("Id: {}\n", id_);
+        fmt::print("Title: {}\n", title_);
+        fmt::print("Artists: {}\n", fmt::join(artists_, ", "));
+        fmt::print("Albums: {}\n", fmt::join(albums_, ", "));
+        fmt::print("Available: {}\n", available_);
     }
 
-    void Track::getSupplement() 
-    {
-        rapidjson::Document document;
-        Request request;
-        string url{"tracks/" + id + "/supplement"};
-        request.makeRequest(url, document);
-        const rapidjson::Value &result = document["result"];
-        assert(result.IsObject());
-
-        supplement = make_shared<Supplement>();
-
-        /// Keys {id: string, lyrics: object}
-        for (auto &key: result.GetObject()) 
-        {
-            if (string(key.name.GetString()) == "lyrics")
-            {
-                for (auto &key1: key.value.GetObject()) 
-                {
-                    if (string(key1.name.GetString()) == "id")
-                        supplement->setId(key1.value.GetInt());
-
-                    if (string(key1.name.GetString()) == "lyrics")
-                        supplement->setLyrics(key1.value.GetString());
-
-                    if (string(key1.name.GetString()) == "fullLyrics")
-                        supplement->setFullLyrics(key1.value.GetString());
-
-                    if (string(key1.name.GetString()) == "textLanguage")
-                        supplement->setTextLanguage(key1.value.GetString());
-
-                    if (string(key1.name.GetString()) == "showTranslation")
-                        supplement->setShowTranslation(key1.value.GetBool());
-
-                    if (string(key1.name.GetString()) == "hasRights")
-                        supplement->setHasRights(key1.value.GetBool());
-                }
-            }
-        }
-    }
-
-    void Track::print() 
-    {
-        std::cout << "Id: " << this->id << "\n";
-        std::cout << "Title: " << this->title << "\n";
-        std::cout << "Artists: ";
-        for (const auto &artist: this->artists)
-            std::cout << artist << "\t";
-        std::cout << "\n";
-        std::cout << "Artists: ";
-        for (auto album: this->albums)
-            std::cout << album << "\t";
-        std::cout << "\n";
-        std::cout << "Available: " << this->available << "\n";
-    }
-
-}
+} // namespace YandexMusic
